Changed Sound Manager to Audio Manager (User Control) added Visuals to Audio Manager
Deleted Sound Manager and replaced with Audio Manager. The latter, as a
user control, can operate autonomously, without laying more code to run
with the grid. This also allows it to have a looping track in the
background, whilst also handling single fire sounds much better. Uses an
MP3 as the Main Game Music, at the moment.

Also, you'll noticed the Visual that the audio manager has (which I
created). It doesn't have to have visuals. I wasn't sure whether we
handle this, or it's something set outside with nsquared. Let me know
whether we want the visuals or not. They can be turned off.