Score up/down callable as a function - Player now earns points for destroying obstacles , loses points for dying
Created a new function to handle score change. To use:
LIVESSCOREREF.ChangeScore(INT, BOOL). The parameters include an int, for
the value to change the score by, and the bool, positive for score going
up, negative for score going down.

Currently called when player is caught in own blast, and when
destructible wall caught in blast.

Scores above 100 were being clipped by the column, so made column for
score wider. It then was appearing very close to the hearts, so made an
additional column (a smaller one), between the hearts, and where the
score appears. This is working much nicer.