Updated the Audio Manager to be Globally Controlled, updates to Main Window, player lives and score, deleted game manager
Have made audio globally available as a static bool (need to protect it)
- though not too dangerous, essentially will just either mean audio is
on or off). Moved Audio off/on to MainWindow, and removed slider as
suggested by Chris.
Updated Lives and Score class, so player gets a message when they have
run out of lives.
Added more obstacles to Grid, to allow further testing of bombs, score
and lives.
Removed Game Manager, as GameTimer will be within Game Manager, as
suggested by Chris. Need to update GameTimer to be a more prototypical
Game Loop, where the 'game' will only run as long as the 'while' loop is
running. This will allow us to place a main title screen and
instructions, without having to create more Windows.