WizardWarz Design Implemented in parts
Have placed a few of Ben's assets in, to give it a stronger feel. We
might need him to update some to be 64px, instead of 32px, as they come
up a bit blurry. New art included is for the hearts, the home base
castle, and the wizard.

I created some additional designs for wizard front and back
(placeholders only, as Ben has actual ones).

Player changes directions (with their image) as they move. The code
should be fine for implementing the animations too, rather than just a
still. It essentially checks the tile between each move (old and new
tile), and depending on the direction, will update the sprite.