Highlight movement path - Incomplete (doesn't remove after)
So far it highlights where the user has dragged, but I want them to all disappear after their finger is raised (which can tie in with the mouse released event nicely). I've got a stinking headache, and can't think of how to remove them, as they are being added to the localgrid, and I can't clear the whole thing, without deleting everything else on the board.