Update to Player Controllers - movement working:
Player Controller update. Players now set correctly to their place on
the board (though still need to be deeper and further across), and they
actually match their lives and score area.
Highlights have been fixed, with players highlighting correctly.
Player image is set more robustly, wherein a string is used (the
resource name) and passed back to a variable, which uses that to set the
player image.
Made all sections which hard coded 64px tileSize to be robust, and use
the GameWindow's Returned TileSize, so everything follows the same size,
and if it ever needs to change, it can be done easily.
Updated a lot of Code with summaries, so that when used other
programmers can see through Intellisense what the functions do and need.

Generally added a lot more comments.